moved sniffer to hijack package
